split index.js into several files + add DataRecipient type
